feat #21: 상대 정보 담을 클래스 분리
- 동행 구인글 상세 조회 및 댓글 목록 반환 시, 작성자의 정보 담기 위함.
